Add some basic styling with Reactstrap (#13)
* refactor pages and components
* add a Header
* style signin form
