+++ components/Header.js
@@ -1,0 +1,39 @@
+import React, { useState } from 'react';
+
+import {
+  Collapse,
+  Navbar,
+  NavbarToggler,
+  NavbarBrand,
+  Nav,
+  NavItem,
+  NavLink
+} from 'reactstrap';
+
+const Header = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const toggle = () => setIsOpen(!isOpen);
+
+  return (
+    <div>
+      <Navbar color="light" light expand="md">
+        <NavbarBrand href="/">MyWordlist</NavbarBrand>
+        <NavbarToggler onClick={toggle} />
+        <Collapse isOpen={isOpen} navbar>
+          <Nav className="mr-auto" navbar>
+            <NavItem>
+              <NavLink href="/signin">Sign In</NavLink>
+            </NavItem>
+            <NavItem>
+              <NavLink href="/mywordlist">My Wordlist</NavLink>
+            </NavItem>
+          </Nav>
+          {/* <NavbarText>Simple Text</NavbarText> */}
+        </Collapse>
+      </Navbar>
+    </div>
+  );
+};
+
+export default Header;
