+++ pages/mywordlist.js
@@ -1,0 +1,6 @@
+import Layout from '../components/Layout';
+import MyWordlist from '../components/MyWordlist';
+
+const MyWordlistView = () => <Layout><MyWordlist /></Layout>;
+
+export default MyWordlistView;
